Tests/LibWeb/Text/expected/selection-drag-past-short-logo-in-row.txt

short logo was selected: true
short logo stayed selected: true
